pyproject.toml
[project]
name = "zs.zflow"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ZackaryW", email = "gitzackw@gmail.com" }
]
dependencies = [
    "pyyaml>=6.0.2",
    "zuu>=7.3.0",
    "croniter>=6.0.0",
    "dateparser>=1.2.1",
    "psutil>=7.0.0",
    "pygetwindow>=0.0.9",
    "click>=8.1.8",
    "watchdog>=6.0.0"
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/zs"]
